pyproject.toml
[build-system]
requires = ["setuptools >= 40.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "informa"
description = "API scraper and more based on rocketry"
dynamic = ["version", "readme"]
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "beautifulsoup4==4.11.1",
	"click==8.1.3",
	"dataclasses-jsonschema==2.16.0",
    "fake-useragent==1.1.1",
    "feedparser==6.0.10",
	"Jinja2==3.1.2",
	"pyyaml==6.0",
    "requests==2.28.1",
	"rocketry==2.5.1",
]

[project.scripts]
informa = "informa.cli:cli"

[tool.setuptools.packages.find]
exclude = ["test*", "venv*"]

[tool.setuptools.package-data]
informa = ["*.txt", "*.md", "LICENSE"]

[tool.setuptools.dynamic]
version = {attr = "informa.__version__"}
readme = {file = ["README.md"]}

[tool.distutils.bdist_wheel]
universal = true
